bug fix parking init demand
git-svn-id: file:///srv/local/data/cleanRepo/trunk@16213 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/wrashid/src/main/java/playground/wrashid/parkingChoice/infrastructure/FlatParkingFormatReaderV1.java
@@ -36,7 +36,7 @@ public class FlatParkingFormatReaderV1 extends MatsimXmlParser implements Matsim
 		String parkingId=atts.getValue("id");
 		double x=new Double(atts.getValue("x"));
 		double y=new Double(atts.getValue("y"));
-		int capacity=new Integer(atts.getValue("capacity"));
+		double capacity=new Double(atts.getValue("capacity"));
 		Coord coord=new CoordImpl(x, y);
 		
 		if (parkingType.equalsIgnoreCase("public")){
